Moved the classes of the addon to another package
